+++ src/components/todoList/TodoListItem.jsx
@@ -1,19 +1,10 @@
-import React, {
-  useContext
-} from 'react'
-import { TodosContext } from '../../context/todo'
+import React from 'react'
 
-export const TodoListItem = (props) => {
-  const { deleteTodo } = useContext(TodosContext)
-  const handleClick = (docId) => {
-    deleteTodo(docId)
-  }
-
+function TodoListItem(props) {
   return (
     <li className="group hover:bg-gray-200 duration-150 rounded-md flex justify-between mt-2 p-2 text-xs cursor-pointer">
       <span>{props.list.title}</span>
       <img
-        onClick={() => handleClick(props.list.docId)}
         src="./icon/close.svg"
         alt="delete todo icon"
         className="w-4 opacity-0 group-hover:opacity-100 duration-150"
@@ -21,3 +12,5 @@ export const TodoListItem = (props) => {
     </li>
   )
 }
+
+export default TodoListItem
