+++ src/context/todo.js
@@ -32,12 +32,8 @@ const TodosProvider = ({ children }) => {
     })
   }, [collection])
 
-  const deleteTodo = useCallback(async (docId) => {
-    await collection.doc(docId).delete()
-  }, [collection])
-
   return (
-    <TodosContext.Provider value={{ todos, addTodo, deleteTodo }}>
+    <TodosContext.Provider value={{ todos, addTodo }}>
       {children}
     </TodosContext.Provider>
   )
